requirements.txt
# Core Dependencies
pandas==2.1.3
numpy==1.26.2

# ML/AI
scikit-learn==1.3.2

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
prometheus-client==0.19.0

# AWS (optional - for real AWS data)
boto3==1.34.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
